List of functions in MarkovChainHammer

MarkovChainHammer.TransitionMatrix.generator — Method

generator(markov_chain; dt=1)

Description

Calculate the generator matrix from a markov chain.

Arguments

markov_chain::AbstractVector: A vector of integers representing the state of a markov chain at each time step.

dt::Real: The time step between each state.

Returns

generator_matrix::Matrix: The generator matrix of the markov chain.

MarkovChainHammer.TransitionMatrix.holding_times — Method

holding_times(markov_chain, number_of_states; dt=1)

Description

Calculate the holding times of a markov chain.

Arguments

markov_chain::AbstractVector: A vector of integers representing the state of a markov chain at each time step.

number_of_states::Integer: The number of states in the markov chain.

dt::Real: The time step of the markov chain.

Returns

holding_times::Vector{Vector{Real}}: A vector of vectors of holding times for each state.

MarkovChainHammer.TransitionMatrix.perron_frobenius — Method

perron_frobenius(markov_chain)

Description

Calculate the perron-frobenius matrix from a markov chain.

Arguments

markov_chain::AbstractVector: A vector of integers representing the state of a markov chain at each time step.

Returns

perron_frobenius_matrix::Matrix: The perron-frobenius matrix of the markov chain.

MarkovChainHammer.TransitionMatrix.symmetric_generator — Method

symmetric_generator(markov_chain, symmetries; dt=1)

Description

Calculate the generator matrix from a markov chain with symmetries.

Arguments

markov_chain::AbstractVector: A vector of integers representing the state of a markov chain at each time step.

symmetries::AbstractVector: A vector of functions that are symmetries of the markov chain.

dt::Real: The time step between each state.

Returns

generator_matrix::Matrix: The generator matrix of the markov chain.

MarkovChainHammer.TransitionMatrix.symmetric_perron_frobenius — Method

symmetric_perron_frobenius(markov_chain, symmetries)

Description

Calculate the perron-frobenius matrix from a markov chain with symmetries.

Arguments

markov_chain::AbstractVector: A vector of integers representing the state of a markov chain at each time step.

symmetries::AbstractVector: A vector of functions that are symmetries of the markov chain.

Returns

perron_frobenius_matrix::Matrix: The perron-frobenius matrix of the markov chain.

MarkovChainHammer.Utils.histogram — Method

function histogram( array; bins=minimum([100, length(array)]), normalization=:uniform, custom_range=false )

Description

Utility function for barplot in GLMakie

Arguments

array; one dimensional sequence of numbers

Keyword Arguments

bins; how many buckets to use, always uniform

normalization; how much to weight each value of array (default 1/length(array))

custom_range; range for uniform bucket